+++ src/components/MenuSider/MenuSider.tsx
@@ -128,6 +128,13 @@ const iconsInActive = {
       }}
     />
   ),
+  promotionNoti: (
+    <GiftOutlined
+      style={{
+        fontSize: "20px",
+      }}
+    />
+  ),
   freebies: (
     <GiftOutlined
       style={{