+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 ## [Version - 0]
 
+## Resolved [v0.6.31] - Release Date: 2023-10-02
+
+- fix promotion file download
+
+## Resolved [v0.6.30] - Release Date: 2023-09-29
+
+- fix permission document in promotion detail page
+
+## Resolved [v0.6.29] - Release Date: 2023-09-29
+
+- enhanced promotion noti setting
+- add Feature: noti promotion menu
+
 ## Resolved [v0.6.28] - Release Date: 2023-09-24
 
 - Improve: Permission